Adding LSTM to pipeline
Adding LSTM back into the environment step and reset.
Input of the LSTM are the new action (0 vector in reset) and the new kge
postion (initial position in reset). This encoded path along with the
textual embedding are used to encode a projected state for the
navigation agent.
This is a similar approach to Salesforce.